13: Move some functionality to another class so it can be tested.
