+++ app/build.gradle
@@ -70,7 +70,7 @@ tasks.named('test') {
     useJUnitPlatform()
 
     jacoco {
-        excludes = []
+        excludes = ['**/javalin/App*']
     }
 }
